+++ crates/bcr-ebill-transport/src/lib.rs
@@ -182,6 +182,7 @@ pub async fn create_transport_service(
         nostr_contact_processor.clone(),
         db_context.nostr_chain_event_store.clone(),
         transport.clone(),
+        db_context.contact_store.clone(),
         get_config().bitcoin_network(),
     ));
 
@@ -291,6 +292,7 @@ pub async fn create_nostr_consumer(
         nostr_contact_processor.clone(),
         db_context.nostr_chain_event_store.clone(),
         transport.clone(),
+        db_context.contact_store.clone(),
         get_config().bitcoin_network(),
     ));
 
@@ -431,6 +433,7 @@ pub async fn create_restore_account_service(
         nostr_contact_processor,
         db_context.nostr_chain_event_store.clone(),
         nostr_client.clone(),
+        db_context.contact_store.clone(),
         config.bitcoin_network(),
     ));