+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.5.14
 
 * When creating a company, or being invited to a company, it's added to the contact book (if it hasn't been there before), but without files
+* When creating an identity, or restoring, our own identity is added to the contact book (if it hasn't been there before), but without files
 
 # 0.5.13